현재 지역 날씨 가져오는 코드

+++ app/src/main/java/com/hk/portfolio/today_weather/data/dto/retrofit/weather/WeatherNowDto.kt
@@ -6,5 +6,5 @@ data class WeatherNowDto(
     val category: String?= "",
     val nx:Double?= null,
     val ny:Double?= null,
-    val obsrValue: Int?= null
+    val obsrValue: Float?= null
 )
